requirements.txt
absl-py==0.10.0
aniso8601==8.0.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.2.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
firebase==3.0.1
firebase-admin==4.4.0
Flask==1.1.2
flask-restplus==0.13.0
gast==0.3.3
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
